More 32/64 bit warnings fixed

+++ Classes/CocoaUtility.mm
@@ -59,7 +59,7 @@
 @implementation NSString(StuartsExtra)
 
 - (NSString *)reversed {
-	int len = [self length];
+	size_t len = [self length];
 	char buf[len + 1];
 	buf[len] = '\0';
 	const char* src = [self cStringUsingEncoding:[NSString defaultCStringEncoding]];
