+++ uae4all_gp2x_0.7.2a/src/include/memory-uae.h
@@ -45,7 +45,7 @@ extern uae_u32 allocated_a3000mem;
 #define kickmem_start 0x00F80000
 
 extern int ersatzkickfile;
-extern int cloanto_rom;
+extern size_t cloanto_rom;
 
 typedef struct {
     /* These ones should be self-explanatory... */
